requirements.txt
aniso8601==8.0.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
geographiclib==1.50
geopy==2.0.0
gmplot==1.4.1
haversine==2.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyproj==2.6.1.post1
pytz==2020.1
radar-python==0.0.4
requests==2.24.0
Shapely==1.7.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
Werkzeug==1.0.1
